// src/coaching/enums/suspendCoacheeByOrganization.enum.ts
export enum suspendCoacheeByOrganization {
  NOT_FOUND_COACHEE = 'NOT_FOUND_COACHEE',
  NOT_OWNER_ORGANIZATION_SUSPEND_COACHEE = 'NOT_OWNER_ORGANIZATION_SUSPEND_COACHEE',
  COACHEE_FROM_ANOTHER_ORGANIZATION = 'COACHEE_FROM_ANOTHER_ORGANIZATION',
  COACHEE_ALREADY_SUSPENDED = 'COACHEE_ALREADY_SUSPENDED',
}