// src/coaching/enums/activateCoacheeByOrganization.enum.ts
export enum activateCoacheeByOrganization {
  NOT_FOUND_COACHEE = 'NOT_FOUND_COACHEE',
  NOT_OWNER_ORGANIZATION_ACTIVATE_COACHEE = 'NOT_OWNER_ORGANIZATION_ACTIVATE_COACHEE',
  COACHEE_FROM_ANOTHER_ORGANIZATION = 'COACHEE_FROM_ANOTHER_ORGANIZATION',
  COACHEE_ALREADY_ACTIVE = 'COACHEE_ALREADY_ACTIVE',
}